added chnages from main

+++ gnc/ekf.h
@@ -48,8 +48,8 @@ private:
     float Wind_alpha = 0.85f;
     float Cp = 0;
     float curr_mass_kg = mass_full; //(kg) Sustainer + Booster, but value changes over time.
-    std::vector<double> starting_gps;    // latitude, longitude, altitude
-    std::vector<double> starting_ecef;   // x, y, z
+    std::vector<float> starting_gps;    // latitude, longitude, altitude
+    std::vector<float> starting_ecef;   // x, y, z
 
     // Eigen::Matrix<float,3,1> gravity = Eigen::Matrix<float,3,1>::Zero();
     KalmanState kalman_state;
@@ -71,10 +71,10 @@ private:
     Eigen::Matrix<float, NUM_CONTROL_INPUTS, 1> u_control;
     
     // GPS reference coordinates
-    double gps_latitude_original = 0.0f;
-    double gps_longitude_original = 0.0f;
-    double gps_latitude_last = 0.0f;
-    double gps_longitude_last = 0.0f;
+    float gps_latitude_original = 0.0f;
+    float gps_longitude_original = 0.0f;
+    float gps_latitude_last = 0.0f;
+    float gps_longitude_last = 0.0f;
 };
